zephyr.bbclass: Add ability to set board
Currently, there is no ability to set the board used by Zephyr.  This
limits the ability to set the board to something other than the machine
name.  By setting up an intermediate variable, ZEPHYR_BOARD, this can be
set by those that know better (with the default to use the machine name
as before).

Signed-off-by: Jon Mason <jon.mason@arm.com>