zephyr-kernel-src.inc: Add backport patch for storage partition
Patch already merged upstream, but after the 2.6 release we are based
on. Taking the backport in here until we can figure out if that can go
to the Zephyr 2.6 LTS branch.

The storage partition itself is needed on the Arduino Nano e.g. for
OpenThread or applications needed a storage space.

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>
Signed-off-by: Naveen Saini <naveen.kumar.saini@intel.com>