west: Add west and python dependencies
We will need west for a few things.

First, west knows about all the boards supported, so for automatic
machine generation, we will need to be able to run west boards.

Secondly, I want to get rid of the dependance on MACHINEOVERRIDES in
machine definition files. One of the things that drives me a bit mad
about how we're doing this is effort replication. Zephyr knows what the
machine is, and modules needed and includes. We shouldn't need to
replicate that effort. Future patches will address that through
automatic machine generation.

Signed-off-by: Eilís Ní Fhlannagáin <elizabeth.flanagan@huawei.com>
Signed-off-by: Naveen Saini <naveen.kumar.saini@intel.com>