recipes-meta: Abuse CMake to create OE machine definitions
Trying to handwrite all the machine defintions in zephyr would be an
enormous task. This commit creates a patch to zephy which abuses cmake
to output machine definitions for OE.

Not all machines can be generated this way and meta will need to be in
COREBASE/meta. It's mainly a prototype tool at this point to be able to
add all the machines this patch can figure out to meta-zephyr-bsp.

The hope is at a future date we can use it to autogenerate all the
machines we can from zephyr and use them in meta-zephyr-bsp.

Signed-off-by: Eilís Ní Fhlannagáin <elizabeth.flanagan@huawei.com>
Signed-off-by: Naveen Saini <naveen.kumar.saini@intel.com>