zephyrtest: Enable use of TESTIMAGE_AUTO
When TESTIMAGE_AUTO is enabled, the do_testimage task is inserted after
do_image_complete and before do_build so that the test suites
automatically run as part of the image build.

However, do_testdata_write is currently constrained to run only before
do_build, so it likely won't execute prior to do_testimage. Change the
"before" constraint to do_testimage do that the testdata is always
generated prior to running the testimage task.

Signed-off-by: Peter Hoyes <Peter.Hoyes@arm.com>
Signed-off-by: Naveen Saini <naveen.kumar.saini@intel.com>