CI: add more targets
The autotest change allows for the list of targets to be built to be
larger.  Add more targets to increase the build coverage of CI.
Unfortunately, there are some platforms that do not build everything.
So, clip those down to those that work for all.  Also, there are a few
targets unique to certain platforms.  So, expand those to have coverage.

Signed-off-by: Jon Mason <jon.mason@arm.com>
Signed-off-by: Naveen Saini <naveen.kumar.saini@intel.com>