zephyr-kernel: Checkout zephyr in subdir to match West usage
When using West, Zephyr is checked out in a "zephyr" subdirectory of the
working directory. At the moment Zephyr is checked out at the root,
meaning a workaround is required to extract the module information.

In order to remove the workaround, change the Zephyr destsuffix in
SRC_URI and ZEPHYR_BASE. As a result of this change, modify patchdir for
the patches and the path to the LICENSE file.

Signed-off-by: Peter Hoyes <Peter.Hoyes@arm.com>
Signed-off-by: Naveen Saini <naveen.kumar.saini@intel.com>