zephyr-core/classes: Remove West-based logic from zephyr.bbclass
For a given release tag, ZEPHYR_MODULES is constant. It is therefore now
populated by the generate-version.py script, the output of which is
stored in the repository, so the build-time logic to populate
ZEPHYR_MODULES is no longer needed.

Remove the do_get_zmods task, but retain the Python dependencies which
are still required by Python scripts in the Zephyr repository that are
trigerred by CMake.

The above means that West is now only required as a host dependency to
run generate-version.py, so remove the West Yocto recipe.

Signed-off-by: Peter Hoyes <Peter.Hoyes@arm.com>
Tested-by: Jon Mason <jon.mason@arm.com>
Signed-off-by: Naveen Saini <naveen.kumar.saini@intel.com>