zephyr-core/zephyr-kernel-tests: remove now passing tests from exclude list
Retested the zephyr kernel tests to verify which ones are now building
and which ones are now passing CI on qemu machines.  They are organized
as follows:
* Any test that fails to be configured is added to a list in
  zephyr-kernel-test.inc
* Any test that fails to compile for all machines is added to a list in
  zephyr-kernel-test.inc
* Any test that fails to compile for a specific machine is added to a
  machine specific list in zephyr-kernel-test.inc
* Any test that fails CI is added to a list in the relevant machine yml
  file

Hopefully by making this way of doing things uniform, it will enable
better testing and CI.

It is worth noting that with this version of zephyr, gen_isr_table and
other tests are now working for a number of machines.

Signed-off-by: Jon Mason <jon.mason@arm.com>
Signed-off-by: Naveen Saini <naveen.kumar.saini@intel.com>